refactor(facility): move the staff visit insert to its own endpoint
POST /api/facility/visits could not ship. Adding a new verb to an existing
legacy (unregistered) route file cannot satisfy the security lints in any PR
ordering: registry-first trips orphan-registry (the file exports no such verb
yet), route-first trips new-route-old-authz, and shipping both together trips
boundary isolation. Filed as #1491.

POST /api/facility/visits/insert has no prior route file, which is exactly the
register-first state orphan-registry already warns for, so the registry entry
(#1492) lands first and this route follows. Behaviour is unchanged — the
handler moves verbatim, and the gate stays ['isSysadmin', 'isBoardMember'].

Also adds the negative-authz coverage the new path now owns in its own right:
a wrong-role rejection case plus its AUTHZ_TESTED entry. The drift guard caught
this — while the handler lived inside facility/visits/route.ts it inherited
that path's existing coverage, and a route whose whole subject boundary is the
role gate should not be relying on a sibling's test.
