+++ checkin-app/src/__tests__/authzRegistry.test.ts
@@ -70,6 +70,7 @@ const AUTHZ_TESTED = new Set<string>([
     'facility/badges',
     'facility/trends',
     'facility/visits',
+    'facility/visits/insert',
     'finance-ops/payments',
     'finance-ops/payments/[id]',
     // POST + GET deny-paths (401 anon / 403 non-board) in s-read/sync/__tests__/route.test.ts,
